+++ src/components/header/style.css
@@ -4,6 +4,7 @@ header h1 {
     font-weight: bold;
     font-size: 50px;
     margin-bottom: 0px;
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
 }
 
 .headerPng {
@@ -24,4 +25,5 @@ header p {
     padding-top: 20px;
     padding-bottom: 20px;
     margin-bottom: 0px;
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
 }
